Move json_utils.rs -> utils/json.rs
