+++ deepwell/src/utils/json.rs
@@ -1,5 +1,5 @@
 /*
- * json_utils.rs
+ * utils/json.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2022 Wikijump Team
