+++ deepwell/src/utils/mod.rs
@@ -22,6 +22,7 @@
 
 mod category;
 mod error;
+mod json;
 mod locale;
 mod string;
 mod tide;
@@ -30,6 +31,7 @@ mod user;
 
 pub use self::category::*;
 pub use self::error::*;
+pub use self::json::*;
 pub use self::locale::*;
 pub use self::string::*;
 pub use self::tide::*;
